+++ apis/idpV1Strc.go
@@ -1,7 +1,9 @@
 package apis
 
 type IDPDataSet struct {
-	Asset IDPAsset
-	Debt  IDPDebt
-	Macro IDPMacro
+	Asset     IDPAsset
+	Debt      IDPDebt
+	Macro     IDPMacro
+	ModelInfo []byte
+	ModelCoef []byte
 }